requirements.txt
azure-functions
pyodbc
dotenv~=0.9.9
python-dotenv~=1.1.0
pymssql~=2.3.4